+++ lib/astarte_rpc/protocol/proto/housekeeping/update_realm.pb.ex
@@ -1,9 +1,21 @@
+defmodule Astarte.RPC.Protocol.Housekeeping.RemoveLimit do
+  @moduledoc false
+
+  use Protobuf, protoc_gen_elixir_version: "0.12.0", syntax: :proto3
+end
+
+defmodule Astarte.RPC.Protocol.Housekeeping.SetLimit do
+  @moduledoc false
+
+  use Protobuf, protoc_gen_elixir_version: "0.12.0", syntax: :proto3
+
+  field :value, 1, type: :int64
+end
+
 defmodule Astarte.RPC.Protocol.Housekeeping.UpdateRealm.DatacenterReplicationFactorsEntry do
   @moduledoc false
 
-  use Protobuf, map: true, protoc_gen_elixir_version: "0.11.0", syntax: :proto3
-
-  def fully_qualified_name, do: "UpdateRealm.DatacenterReplicationFactorsEntry"
+  use Protobuf, map: true, protoc_gen_elixir_version: "0.12.0", syntax: :proto3
 
   field :key, 1, type: :string
   field :value, 2, type: :int32
@@ -12,9 +24,9 @@ end
 defmodule Astarte.RPC.Protocol.Housekeeping.UpdateRealm do
   @moduledoc false
 
-  use Protobuf, protoc_gen_elixir_version: "0.11.0", syntax: :proto3
+  use Protobuf, protoc_gen_elixir_version: "0.12.0", syntax: :proto3
 
-  def fully_qualified_name, do: "UpdateRealm"
+  oneof :device_registration_limit, 0
 
   field :realm, 1, proto3_optional: true, type: :string
   field :jwt_public_key_pem, 3, proto3_optional: true, type: :string, json_name: "jwtPublicKeyPem"
@@ -26,7 +38,7 @@ defmodule Astarte.RPC.Protocol.Housekeeping.UpdateRealm do
 
   field :replication_class, 5,
     proto3_optional: true,
-    type: Astarte.RPC.Protocol.Housekeeping.ReplicationClass,
+    type: ReplicationClass,
     json_name: "replicationClass",
     enum: true
 
@@ -35,4 +47,14 @@ defmodule Astarte.RPC.Protocol.Housekeeping.UpdateRealm do
     type: Astarte.RPC.Protocol.Housekeeping.UpdateRealm.DatacenterReplicationFactorsEntry,
     json_name: "datacenterReplicationFactors",
     map: true
+
+  field :set_limit, 7,
+    type: Astarte.RPC.Protocol.Housekeeping.SetLimit,
+    json_name: "setLimit",
+    oneof: 0
+
+  field :remove_limit, 8,
+    type: Astarte.RPC.Protocol.Housekeeping.RemoveLimit,
+    json_name: "removeLimit",
+    oneof: 0
 end