+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Add support for device deletion.
 - Add support to realm update.
+- Add `device_registration_limit` field to `CreateRealm`, `GetRealmReply`
+  and `UpdateRealm`.
 
 ## [1.1.0] - 2023-06-20